:+1: TSに修正

+++ components/LikeButton.vue
@@ -15,7 +15,7 @@
   </button>
 </template>
 
-<script setup>
+<script setup lang="ts">
 const supabase = useSupabaseClient();
 const router = useRouter();
 const props = defineProps({
@@ -36,18 +36,20 @@ const countLike = async () => {
       .select("*")
       .eq("userId", userId)
       .eq("articleId", articleId);
-    if (!confirmation.data[0]) {
-      // Likeテーブルにデータを挿入
-      await supabase.from("like").insert({ userId, articleId });
-      router.go();
-    } else {
-      //いいね数を削除する
-      await supabase
-        .from("like")
-        .delete()
-        .eq("userId", userId)
-        .eq("articleId", articleId);
-      router.go();
+    if (!(confirmation.data === null)) {
+      if (!confirmation.data[0]) {
+        // Likeテーブルにデータを挿入
+        await supabase.from("like").insert({ userId, articleId });
+        location.reload();
+      } else {
+        //いいね数を削除する
+        await supabase
+          .from("like")
+          .delete()
+          .eq("userId", userId)
+          .eq("articleId", articleId);
+        location.reload();
+      }
     }
     //likeテーブルから該当する記事のいいね数を取得
     const { data, error } = await supabase
