+++ components/RecommendButton.vue
@@ -15,7 +15,7 @@
   </button>
 </template>
 
-<script setup>
+<script setup lang="ts">
 const supabase = useSupabaseClient();
 const router = useRouter();
 const props = defineProps({
@@ -36,20 +36,21 @@ const countRecommend = async () => {
       .select("*")
       .eq("userId", userId)
       .eq("articleId", articleId);
-
-    if (!confirmation.data[0]) {
-      // recommendテーブルにデータを挿入
-      console.log(userId, articleId);
-      await supabase.from("recommend").insert({ userId, articleId });
-      router.go();
-    } else {
-      // //おすすめ数を削除する
-      await supabase
-        .from("recommend")
-        .delete()
-        .eq("userId", userId)
-        .eq("articleId", articleId);
-      router.go();
+    if (!(confirmation.data === null)) {
+      if (!confirmation.data[0]) {
+        // recommendテーブルにデータを挿入
+        console.log(userId, articleId);
+        await supabase.from("recommend").insert({ userId, articleId });
+        location.reload();
+      } else {
+        // //おすすめ数を削除する
+        await supabase
+          .from("recommend")
+          .delete()
+          .eq("userId", userId)
+          .eq("articleId", articleId);
+        location.reload();
+      }
     }
     //recommendテーブルから該当する記事のおすすめ数を取得
     const { data, error } = await supabase
